add test cases for the classes and fixed bugs
