+++ src/classicalGuitar.js
@@ -1,0 +1,10 @@
+import AcousticGuitar from './acousticGuitar';
+
+export default class ClassicalGuitar extends AcousticGuitar {
+    constructor(name, origin, numStrings, stringMaterial,
+                isElectricallyOperated, soundProjectionMeans) {
+                    super(name, origin, numStrings, stringMaterial,
+                        isElectricallyOperated, soundProjectionMeans);
+                }
+        
+}
